style: Spark 프레임워크에서 사용하던 클래스 제거
- 클래스 이름 변경
