+++ src/main/java/wooteco/chess/ConsoleApplication.java
@@ -4,7 +4,6 @@ import wooteco.chess.controller.ConsoleController;
 
 public class ConsoleApplication {
 	public static void main(String[] args) {
-		ConsoleController consoleController = new ConsoleController();
-		consoleController.run();
+		new ConsoleController().run();
 	}
 }
