--- src/main/java/wooteco/chess/dao/Connectable.java
@@ -1,9 +1,0 @@
-package wooteco.chess.dao;
-
-import java.sql.Connection;
-
-public interface Connectable {
-	Connection connect();
-
-	void close(Connection con);
-}
