+++ src/main/java/wooteco/chess/service/ChessGameService.java
@@ -1,81 +1,46 @@
 package wooteco.chess.service;
 
-import java.sql.SQLException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import wooteco.chess.dao.RoomDao;
 import wooteco.chess.domain.BoardConverter;
 import wooteco.chess.domain.ChessGame;
 import wooteco.chess.domain.FinishFlag;
 import wooteco.chess.domain.Side;
 import wooteco.chess.domain.position.Position;
 import wooteco.chess.dto.ChessGameDto;
-import wooteco.chess.dto.RoomDto;
+import wooteco.chess.dto.PieceMoveDto;
+import wooteco.chess.dto.RoomNameDto;
+import wooteco.chess.repository.Room;
+import wooteco.chess.repository.RoomRepository;
 
 @Service
 public class ChessGameService {
-	private final RoomDao roomDao;
+	private final RoomRepository roomRepository;
 
-	public ChessGameService(RoomDao roomDao) {
-		this.roomDao = roomDao;
+	public ChessGameService(RoomRepository roomRepository) {
+		this.roomRepository = roomRepository;
 	}
 
-	public ChessGameDto create(String roomName) throws SQLException {
-		validateDuplicated(roomName);
-		ChessGame chessGame = ChessGame.start();
-		roomDao.addRoom(roomName, chessGame);
-		return ChessGameDto.of(roomName, chessGame);
-	}
-
-	private void validateDuplicated(String roomName) throws SQLException {
-		if (isPresentRoom(roomName)) {
-			throw new IllegalArgumentException("입력한 방이 이미 존재합니다.");
-		}
-	}
-
-	private boolean isPresentRoom(String roomName) throws SQLException {
-		return roomDao.findByRoomName(roomName, "board").isPresent();
-	}
-
-	public ChessGameDto move(String roomName, String source, String target) throws SQLException {
-		ChessGameDto chessGameDto = load(roomName);
-		ChessGame chessGame = new ChessGame(BoardConverter.convertToBoard(chessGameDto.getBoard()),
-				Side.valueOf(chessGameDto.getTurn()));
-
-		chessGame.move(new Position(source), new Position(target));
-		roomDao.updateRoom(roomName, chessGame);
-		return ChessGameDto.of(roomName, chessGame);
-	}
-
-	public List<RoomDto> findAllRooms() throws SQLException {
-		List<String> rooms = roomDao.findAll();
+	public List<RoomNameDto> findAllRooms() {
+		List<Room> rooms = roomRepository.findAll();
 		return rooms.stream()
-				.map(RoomDto::new)
+				.map(Room::getRoomName)
+				.map(RoomNameDto::new)
 				.collect(Collectors.toList());
 	}
 
-	public ChessGameDto load(String roomName) throws SQLException {
-		validatePresent(roomName);
-		String finishFlag = roomDao.findByRoomName(roomName, "finish_flag")
-				.orElseThrow(NoSuchElementException::new);
-		validateFinish(finishFlag);
+	public ChessGameDto load(String roomName) {
+		Room room = roomRepository.findByRoomName(roomName).orElseThrow(NoSuchElementException::new);
+		validateFinish(room.getFinishFlag());
 
-		String board = roomDao.findByRoomName(roomName, "board")
-				.orElseThrow(NoSuchElementException::new);
-		String turn = roomDao.findByRoomName(roomName, "turn")
-				.orElseThrow(NoSuchElementException::new);
-		ChessGame chessGame = new ChessGame(BoardConverter.convertToBoard(board), Side.valueOf(turn));
+		String board = room.getBoard();
+		Side turn = room.getTurn();
+		ChessGame chessGame = new ChessGame(BoardConverter.convertToBoard(board), turn);
 		return ChessGameDto.of(roomName, chessGame);
-	}
-
-	private void validatePresent(String roomName) throws SQLException {
-		if (!isPresentRoom(roomName)) {
-			throw new IllegalArgumentException("입력한 방이 없습니다.");
-		}
 	}
 
 	private void validateFinish(String finishFlag) {
@@ -84,9 +49,41 @@ public class ChessGameService {
 		}
 	}
 
-	public ChessGameDto restart(String roomName) throws SQLException {
+	public ChessGameDto create(String roomName) {
+		validateDuplicated(roomName);
 		ChessGame chessGame = ChessGame.start();
-		roomDao.updateRoom(roomName, chessGame);
+		String board = BoardConverter.convertToString(chessGame.getBoard());
+		roomRepository.save(new Room(roomName, board, Side.WHITE, FinishFlag.GOING.getSymbol()));
 		return ChessGameDto.of(roomName, chessGame);
+	}
+
+	private void validateDuplicated(String roomName) {
+		if (roomRepository.findByRoomName(roomName).isPresent()) {
+			throw new IllegalArgumentException("입력한 방이 이미 존재합니다.");
+		}
+	}
+
+	public Room move(PieceMoveDto pieceMoveDto) {
+		String roomName = pieceMoveDto.getRoomName();
+		Room room = roomRepository.findByRoomName(roomName).orElseThrow(NoSuchElementException::new);
+		ChessGame chessGame = new ChessGame(BoardConverter.convertToBoard(room.getBoard()),
+				room.getTurn());
+
+		Position source = new Position(pieceMoveDto.getSource());
+		Position target = new Position(pieceMoveDto.getTarget());
+		chessGame.move(source, target);
+
+		room.update(BoardConverter.convertToString(chessGame.getBoard()), chessGame.getTurn(),
+				FinishFlag.of(chessGame.isEnd()).getSymbol());
+		return roomRepository.save(room);
+	}
+
+	public ChessGameDto restart(String name) {
+		Room room = roomRepository.findByRoomName(name).orElseThrow(NoSuchElementException::new);
+		ChessGame chessGame = ChessGame.start();
+		room.update(BoardConverter.convertToString(chessGame.getBoard()), chessGame.getTurn(),
+				FinishFlag.of(chessGame.isEnd()).getSymbol());
+		roomRepository.save(room);
+		return ChessGameDto.of(name, chessGame);
 	}
 }
